stop polyfilling fetch Response
