+++ realm-server/package.json
@@ -9,16 +9,12 @@
   },
   "devDependencies": {
     "@cardstack/runtime-common": "1.0.0",
-    "@types/node-fetch": "^2.6.2",
     "@types/qunit": "^2.11.3",
     "@types/supertest": "^2.0.12",
     "@types/qs": "^6.9.7",
     "@types/tmp": "^0.2.3",
     "@types/yargs": "^17.0.10",
     "fs-extra": "^10.1.0",
-    "koa": "^2.13.4",
-    "koa-bodyparser": "^4.3.0",
-    "node-fetch": "^2.6.7",
     "qs": "^6.10.5",
     "qunit": "^2.18.0",
     "supertest": "^6.2.4",
