+++ runtime-common/stream.ts
@@ -1,0 +1,18 @@
+export async function streamToText(
+  stream: ReadableStream<Uint8Array>
+): Promise<string> {
+  let decoder = new TextDecoder();
+  let pieces: string[] = [];
+  let reader = stream.getReader();
+  while (true) {
+    let { done, value } = await reader.read();
+    if (done) {
+      pieces.push(decoder.decode(undefined, { stream: false }));
+      break;
+    }
+    if (value) {
+      pieces.push(decoder.decode(value, { stream: true }));
+    }
+  }
+  return pieces.join("");
+}
